Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> - I could easily end up with multiple popups open. I think the menu should close as soon as the mouse moves out of it
> - I think it would be nicer if the menu opened centered over the character it is a zoom of
> - When the mouse is over the menu, the menu item is selected, which looks strange
> 
> I think using a BToolTip rather than a BPopupMenu would make more sense overall?

I am only able to get multiple popups if I click really rapidly, but I do see that is also an issue with the existing right-click popup.  I can try to fix that while I'm at it.

The only reason I used the BPopupMenu is because there was already an existing path for the right-click so it was pretty trivial to reuse it for this.  I'll see if a BToolTip behaves better.  I couldn't decide if the 'selecting item' behavior was bad or good, so I'm glad to get more feedback.

Patch-set: 3
Reviewer: Gerrit User 1000217 <1000217@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>