Icon-O-Matic: check number of items when exporting
Silently saving only part of the data is not nice and confusing.

The error message just says "Out of range" now however, which is not
really great either. We could improve the error reporting, and, as
mentionned in the ticket, also warn in the UI when the limit is reached
(a marker near the shape/path/fill lists maybe?) so the user knows they
need to simplify their picture.

Part of #13978