Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Wouldn't simply checking for IsLocalFile be enough?

One would think so, but I have just experimented with this and it doesn't seem to be the case.  Looking at the logic on `IsLocalFile`, but seems like this method is perhaps checking to see if the package is a local file and it is not installed.

Patch-set: 1
Reviewer: Gerrit User 1000018 <1000018@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>