Update patch set 1
Patch Set 1:

> md5.h is now c++ only, so this is a possibility only for C++ code.

Yes, but this was never meant to be part of the pblic API anyway, it's used internally in the HTTP authentication code and in the APE reader (both C++ code). Previously the APE reader had an independant implementation and the MD5 one was in libnetapi. I moved it to libshared but forgot about the namespacing when doing so (a few weeks ago). It's unlikely that this is already used by something else.

I don't think we want to provide a C API for crypto/hashing functions? Or if we do, we should look at the "Crypto Kit" changes that are somewhere in Gerrit.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>