Update patch set 3
Patch Set 3:

I tried BToolTip.  

BToolTip is private, not well-documented, and incomplete.  Many methods are stubs or have "//TODO!" in them.  The docs have a big red warning that says "Use at your own risk! This class is private and may change before release."  It really seems like the preferred way to use tooltips is via the BView::SetToolTip method, not working directly with BToolTips (and the subclasses thereof).

Despite that, it behaves ok as long as the pointer is not inside the tooltip (which happens when the tooltip is on top of the character), at which point it won't stay open for more than a split second.  I tried SetSticky(true) but it didn't help.  Now, this could trigger a "Finish implementing BToolTip/BToolTipManager" ticket, but then I'd remove the "easy" tag from #2832.

Am I missing something?  If not, I think BPopupMenu is likely the better route in the current environment. 

Regarding the multiple popup issue, that also occurs for the right-click menu.  I tried making the popup menu a member of the class and reusing it, but could not find a safe (aka non-crashing) way to reuse the menu.  Is there any way to force-dismiss a BPopupMenu?  

Thanks, I appreciate any thoughts you have.

Patch-set: 3
Reviewer: Gerrit User 1000217 <1000217@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>