Update patch set 1
Patch Set 1:

After some experimentation... I think `IsLocalFile()` here is taken to mean "just a regular HPKG file on disk and not part of the installed set of packages".  It is possible that a `PackageInfo` can be in a state of `ACTIVATED` and also not have a value for `fLocalFilePath`; this is the case for installed packages on my test system for example.

I suspect that investigating this in-depth will take some time.  For this reason, if you are not happy with the solution as-is, I will permanently enable the tab and myself or somebody else can come back to it at a later date.

Patch-set: 1
Reviewer: Gerrit User 1000018 <1000018@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>