Update patch set 2
Patch Set 2:

> Patch Set 1: Code-Review-1
> Can't we handle the "full" by overpainting, instead of adding an argument?
> And, is the tab position argument really necessary?

There are two questionable changes in this patch set, both to get around
using dynamic_cast in BeControlLook, one for buttons, and one for tabs.

The questionable button change is to add a B_IS_BUTTON flag to BControlLook.
I need to know if I'm a button in DrawLabel to invert the button color.
The B_ACTIVATED flag is not enough because checked checkboxes and selected
radio buttons also have the B_ACTIVATED flags set but we don't want to
invert their label text color, only buttons.

The questionable tab change is adding the tab_position and full parameters
to DrawActiveTab() and DrawInActiveTab() in BControlLook. I need to know
two pieces of information in BeControlLook::DrawInactiveTab(): is the tab
first and is it the tab before the selected tab?

I could overpaint in BeControlLook but I'd have to overwrite parts of the
previous tab when drawing the selected tab, they don't overlap 100%. If I
were drawing in BTab::DrawTab(), this info would come from the the
tab_position and full parameters. The full param is a bit specific to how
BeOS draws tabs.

A better API might be to pass the current tab index and the selected tab
index. This would allow you to determine if you were first tab, the
selected tab, or the tab before the selected tab. However it would not
allow you to know if you were the last tab without more information
(the tab count.) This information is not needed by BeControlLook and
isn't available in the tab_position and full parameters either (although
tab_position could be extended to add it.)

Patch-set: 2
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>