Update patch set 4
Patch Set 4:

> Yeah, those flags need better names at least, but I'm still not sure why they are needed at all, with overpainting... Can you provide a visual explanation?

On Haiku, a clicked button has a dark grey background but the label remains black. On BeOS, the button background is black and the label is white (in fact BeOS probably uses B_OP_INVERT here to avoid redrawing the text).

So we need a way to change the text color specifically for selected buttons.

I would really prefer a separate DrawButtonLabel method rather than hacking with flags.

Patch-set: 4