Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> Poll format is wrong, average value has no meaning. Probably there are 2 group of users: users that use small icons and users that use large icons (sometimes also hide captions).

Yes, I regret using the wrong poll format back then. But the mean of >2.5 suggests, that fewer use a setting of "1" (=16px). Defaulting to "2" seems reasonable to me.
Mind you, this is just the default. Many of us customize our settings (I use "3", for example, and bunch of other options that default to not-set). So we shouldn't bias this to our own usage, but to the majority of the target users.

Patch-set: 2
Reviewer: Gerrit User 1000004 <1000004@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>