Update patch set 2
Patch Set 2: Code-Review-1

But the point is here we don't actually know what the majority of the users want. Due to how the poll was constructed, we could in fact be choosing an option here that everyone hates, if e.g. a lot of people voted for 16px, and a lot voted for 48 or 64px, but they all think that 24px is just wrong. As X512 also pointed out, this setting is closely tied to the display names / don't display names option. And even further, the option does not take HiDPI and font sizes into account, which it should.

IMHO, what we should do is:
 * Make the icon size setting scale based on system font size, i.e. this can be used to increase it further if needed.
 * Run a new poll, this time combining the icon size and display names options and showing who is voting for what amount.
 * Then decide how to change the default.

Additionally, I think we should actually have the "tray" icons change size along with the app icons. At least it seems annoying there is no other way to get the desktray icons to be larger...

Patch-set: 2
Reviewer: Gerrit User 1000011 <1000011@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-1