Update patch set 1
Patch Set 1:

AFAIKS the serialised `BMessage` here is always coming from a block of data embedded in the HPKR file.  If the base url is embedded in the HPKR then it might be handy for informational purposes, but actually it is not much use with respect to mirrors because the same HPKR will be mirrored at different base urls.  I think when the repository is configured on the Haiku computer, the class `ActivateRepositoryConfigJob` is employed and at line 65, it will use the base URL that was provided for by the user.  I could be missing something here, but this is my understanding.

Patch-set: 1
Reviewer: Gerrit User 1000018 <1000018@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>