Update patch set 2
Patch Set 2:

What is the reasoning behind a drop to just a microsecond?  That's so fast the messaging system would not be able to keep up and I'd expect flooding to be a real issue.

I don't think any faster than 250 us would be usable on even the fastest machine.

Patch-set: 2
CC: Gerrit User 1000085 <1000085@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>