Update patch set 1
Patch Set 1:

I'm not sure if this is the right way to go about it. Ideally BPicture wouldn't rely at all on the view state.

There are also interesting problems if someone records a pictures with unbalanced PushState/PopState commands, but I'm not sure what we should do in that case. I can imagine a valid use case where one would record a "setup" picture, used to put the view in a specific (complex) state and a "teardown" picture to undo the effect and restore the view to the original state.

In this use case, recording of such pictures would also leave the view state changed, which I think is unexpected?

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>