Update patch set 1
Patch Set 1: Code-Review+2

In that case shouldn't the BaseURL from the repo be ignored altogether? Was it an error to store an URL there after all? Shouldn't there be only an identifier?

But ok, I guess for the time being this fixes it and we can go about continuing the cleanup in later beta releases...

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+2