Update patch set 1
Patch Set 1:

> Patch Set 1:
> > I'm not sure the current way is actually cleaner; there are still lots of things done in DispatchMessage() that don't really look any different to me. While it wasn't perfect before, at least it was consistent. It's not anymore.
> > 
> > You can always install a message filter on the looper if you actually need access to the source messages for whatever reason.
> 
> BWindow::DispatchMessage messages are related to window (B_WINDOW_ACTIVATED etc.) so it is consistent. 

That part is indeed consistent, I missed that the remaining items were all targeted to the window. However, it's still not logically consistent, as why does DispatchMessage() for BWindow does the same work as MessageReceived() in BView? Why not always MessageReceived()? What's the purpose of DispatchMessage() in the first place, besides calling MessageReceived() AND calling the handler specific hook methods like FrameMoved() -- or MouseMoved()?
It's its only purpose. Just not anymore because now BView::MessageReceived() does the same work.

If that's the better way to deal with this, we should just abandon DispatchMessage() completely. I think the rationale behind this also was to hide the messaging details from MessageReceived(). But is that really necessary? And what's the gain? That's the main question we'll have to answer. A clean solution could then be derived from there.

> Message filters are unreliable solution because multiple message filters can interfere with each other and depend on order.

It's exactly as reliable as overriding MessageReceived(). You always know the target handler in your message filter. Multiple filters can interfere with each other in exactly the same way as several MessageReceived() implementations can interfere with each other. Where is the difference?

Even if you revert it to the BeOS version, what is the problem of overriding MouseMoved() and friends as well? It's not like it will hide anything from you, it will just distribute it differently.

Patch-set: 1
Reviewer: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>