Update patch set 1
Patch Set 1: Code-Review+2

> Wouldn't both numbers be interesting to know and report?

It's indeed misleading if reported as "swap" (since nothing is swapped). A measurement of "commited memory" would indeed be useful, because that's the main limitation to using the memory in our case, and the point where allocations start to fail. But we should have a better name for it.

I think we can merge this (the new value makes more sense in terms of swap) and we can reintroduce this one elsewhere (it's not like it is very complicated to get) if desired.

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+2