Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Wouldn't both numbers be interesting to know and report?

Indeed, but it shouldn't be the reserved swap size that is reported, but the overall reserved space, as Adrien already said.

The reservation strategy currently is to reserve as much as possible in swap and then reserve physical memory. So the number would be bounded by the swap file size and often be an incomplete view of the actual reservations, which is rather confusing. The split of the reservations is IMO an implementation detail that doesn't really need to be made visible.

Patch-set: 1
Reviewer: Gerrit User 1000017 <1000017@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>