Update patch set 4
Patch Set 4:

> Oh, yes, that was poorly worded :-) I meant that I tried to shutdown Haiku every day in the past days after updating to a Haiku revision with this change included, but never got it right on first try so far. I always don't really look, and choose the second item in the list, which is now reboot :-)

I'm starting to wonder if the second item is actually easier to reach in the default setup: it is a nice diagonal movement from the leaf menu button, to the shutdown menu, to the second item. While reaching the first item needs a curved trajectory.

So the new ordering makes sense for new users who will be reading the menu in-order, but it indeed makes shutdown less easy to reach when you know the layout?

Patch-set: 4