libnetapi/HttpResult: correctly convert length
`atoi()` maximum value is `LONG_MAX` which is smaller than `SIZE_MAX`.
This cause any Content-Length > LONG_MAX to be represented in an
erroneous value. This will also happen to any invalid Content-Length
header.

This change uses strtoul() for the extended range (should be the same as
size_t range), and combined with proper error checking to determine
whether the received Content-Length is a valid value. Returns 0 if the
data is invalid.

Change-Id: If28c4c3b8430ed83dd0f600030ec8949cf7e0051