libnetapi/HttpResult: improve Length() algorithm
`atoi()` maximum value is `LONG_MAX` which is smaller than `SIZE_MAX`.
This cause any Content-Length > LONG_MAX to be represented in an
erroneous value. This will also happen to any invalid Content-Length
header.

This change uses strtoul() for the extended range (should be the same as
size_t range), and combined with proper error checking to determine
whether the received Content-Length is a valid and/or representable
value. Returns 0 if the data is invalid or can not be represented.

Some shortcomings about the current implementation issues are also
noted accordingly. They will be solved in later patches.

Change-Id: If28c4c3b8430ed83dd0f600030ec8949cf7e0051