Update patch set 1
Patch Set 1:

Shouldn't we just add some keyboard shortcuts instead? Like, pressing escape to restore some default value, and shortcuts to increase/decrease brightness?

Also, I'm not sure this will work in all cases. The brightness is sent as a float to the driver (in the range 0..1) and there is no guarantee about the first value that results in the backlight not being completely off. So, this should instead be done inside the driver, which knows more about that.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>