Update patch set 1
Patch Set 1:

> Fair enough. Though the slider now providing values between 0 and 255, it would not limit the lowest possible brightness appreciatably if it changed to between 1 and 255. This would take care of drivers that weren't fixed yet WRT not turning off the screen for brightness=0 at least.

Currently there is one single driver to fix. And that also avoids mishaps with the command line version and anything that uses the API. Sounds much better to me than doing it in the screen preferences.

It's not difficult, either, just move the code on the accelerant side:

https://git.haiku-os.org/haiku/tree/src/add-ons/accelerants/intel_extreme/mode.cpp?id=8ddec194c646653118cc7b7e2b820f4f5584ae9e#n684

duty = std::min(duty, 2); here should do it.

Patch-set: 1