Update patch set 1
Patch Set 1:

> Now: What range of values can "duty" have? Why limit it to below "2"?

The range is from 0 to the value of "period" which we don't change (it's set by the BIOS or the hardware and there is no reason to mess with it). The value must be even because odd values don't work on some Atom CPUs (I don't know why, we just found it out while testing).

So, 2 is the minimal value we can use.

> Including <algorithm>, I get errors when compiling...

What's the error?

Patch-set: 1