Update patch set 1
Patch Set 1:

We should really set up abi_compliance_checker to run easily for this kind of change.

Besides the ABI there's also the risk of different behavior. Possibly not in these functions, but in other parts of netresolv. Things that come to mind:
- Patches to move the settings from /etc to /system/settings/network (for resolv.conf, etc)
- Patches to have sane default values when said settings files are missing
- Behavior of getaddrinfo which uses getifaddrs to decide if it should return IPv6 addresses
- All kind of subtleties with the __ weak aliases and/or defines for all the functions

Please wait a few hours so I can do an in-depth review here if possible :)

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>