Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Implementation approach is wrong. Resize border should be implemented in same way as cursor and dragged bitmap in HWInterface. Direct XOR drawing will not work because windows are painted asynchronously and can erase border causing artifacts.

Indeed this is wrong, but it is very close to being right. My patch currently is based off of this old work and I am getting a border that will independently resize of the window contents. In addition, a decorator can draw its own special “outline border” if wanted (and I made the Default Decorator have a very modern one). As expected however, it is very hard to get the border to draw on top of the window contents. Trying to hide the TopView() of the Window or setting the B_TRANSPARENT_BACKGROUND flag and B_TRANSPARENT_COLOR as the view color both lead to continuous smearing. As looncraz mentioned, looking into how clipping and dirty regions are handled seem like the best solution. After all, I don’t believe that the app_server knows how to make a Window “transparent” at the moment. I just need to continue looking around the cursor handling to see how to do this..

Patch-set: 1
CC: Gerrit User 1000052 <1000052@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>