WIP: Updated Mouse input server to handle multiple mouse settings
Able to store the settings of different mouse in Mouse_settings.

Problem 1:
Now, The GUI changes according to the settings changed for each mouse. But responding as same we made change int the settings for the recent device.

Problem 2:
Suppose I have aatched the 6 button mouse in USB port 1 and set this mouse to 5 button mouse. Then, I diattcahed it and attche the 8 button mouse to the same port (USB port 1).
Again, it shows the type (5 button mouse ) of 8 button mouse.
This is because we are taking the name of the mouse (that is generated input server file) not the serial number of mouse, we are not taking any identification of mouse device.


Change-Id: I37dd27cfa3ac55c6956cc4fb7fc8fabc85a59448