WIP: Updated Mouse input server to handle multiple mouse settings
Able to store the settings of different mouse in Mouse_settings.

Garbage value issue is now resolved. Now, the settings for each mouse are saved and retrieved but, the GUI shows the same settings again for multiple mouse. From the commandline I debug the settings are stored differently for each mouse.

Change-Id: I37dd27cfa3ac55c6956cc4fb7fc8fabc85a59448