Mouse Preferences: store settings for multiple mouses.
Use a BMessage for the mouse settings. Store an array of mouse_setting
and an array of names for the corresponding mouses.

The input preferences can load the settings in both the old and new
format. If in the old format, the settings are applied to all mouses.

In this commit, input_server is not modified yet to use the new settings
format, leading to unusable mouse.

Change-Id: I37dd27cfa3ac55c6956cc4fb7fc8fabc85a59448