WIP: Updated Mouse input server to handle multiple mouse settings
Able to store the settings of different mouse in Mouse_settings.

Getting the garbage value while saving the settings. Retrieve stores
the correct settings but save not.
Steps to produce the code:
- Go to the directory (haiku/src/preferences/input)
- run jam
- remove the file (rm ~/config/settings/Mouse_settings)
- run ../../../generated/objects/haiku/x86_64/release/preferences/input/Input
- close the preferences from GUI
- run it again using the above command.

First you see when you open the input preferences the retrieve logs not present because,
there is no Mouse_settings present. Now, after running the input preferences again you see
the earlier save settings and the new retrieved mouse settings are same.
But, when input preferences saved the file Mouse_settings, it prints the garbage values.

I also check, in RetrieveSettings() and the SaveSettings() methods, MouseSettings objects
have same address.

Can anybody help me to fix this?

Here are the logs:
https://paste.ubuntu.com/p/bF6Vz4mHVz/

Change-Id: I37dd27cfa3ac55c6956cc4fb7fc8fabc85a59448