Update patch set 9
Patch Set 9:

(17 comments)

Hi,

There are still a lot of code formatting problems. Please spend more time reviewing the coding style as it is quite annoying to list them all, and the more interesting points about the actual functionality of the code tends to be lost in these reviews (because the coding style problems are distracting us from it and making the code less easy to read).

Additionally there are some comments about naming. That is also an important part to making the code easier to read and understand, which, to me, is as important as getting it to compile and run, especially in projects like Haiku which are maintained by a somewhat large team and for a very long time.

Patch-set: 9
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>