Update patch set 2
Patch Set 2: Code-Review+1

(6 comments)

Hi!

The code in AggCompOpAdapter is templated to accept any type as the color, but then hardcodes the size to 4 bytes. I think it would be nice to make it adjustable, even if we are unlikely to use non-32bit data for now. I think the compiler should optimize things and result in the same code.

The documentation of the blending modes would probably be easier to understand with illustrations (screenshots from the test app, maybe?)

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+1