Update patch set 2
Patch Set 2: Code-Review+2

I don't see why we should avoid the STL at all costs. It does its job appropriately. It's just that it takes a different approach to things.

BList is a type-unsafe container (it just stores pointers without caring what they point to). BObjectList is a template-based wrapper around it that adds the type safety. As a result, the code in BList is compiled only once, and is reused by all instanciations of BObjectList.

On the other hand, STL containers are templated and generate all the code in every instanciation. This results in more generated code, but more importantly the objects are stored directly inside the container (the vector) rather than having pointers to them, removing one level of indirection when accessing the objects, but creating other problems: the objects must be constructed or moved into the list, or you must build a vector of pointers yourself if that's what you want.

The STL approach has a lot of flexibility: you can have a vector of objects, a vector of owning pointers, a vector of not-owning pointers, ...

Sometimes such things are needed, and it makes sense to use the STL.

What I'd avoid is using STL containers in public API, because it's less readable there. But implementing a container in terms of STL and then exposing it with a plain API is fine, I'd say.

Anyway, using BObjectList here looks just as fine.

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+2