Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Upon further thought, I think we may want to call this MemoryRingIO.

Sure.

> I took a quick look over the code, the only thing I wonder about is maybe delegating to memcpy or similar as that may be better optimized that just a naive byte-by-byte copy loop.

I think gcc can already vectorize a simple loop like this, but I might be wrong.

I do plan to make optimizations that would allow for free space to be represented as a contiguous buffer, which would allow us to just write to it directly. The details are a bit messy so I left it out for now. The performance is good enough as is :)

Patch-set: 1
Reviewer: Gerrit User 1000035 <1000035@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>