Update patch set 3
Patch Set 3:

(6 comments)

Shouldn't this be in the BPrivate namespace initially?

I would say it is not complete yet, depending on the use cases.

For example when we integrated http support in the media kit, we found out it is useful to report the full size of the data to the media kit, even if not all of it is downloaded and immediately available. So if possible, I would not make this part of the public API yet.

Patch-set: 3
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>