Update patch set 1
Patch Set 1:

I would prefer if we kept the "cpu" name everywhere.

Why is there confusion with the kernel sources here? Sounds like SEARCH_SOURCES was applied too liberally?

Usually what we do is have a separate .o or .a built from kernel objects we want to reuse, so that we can set SEARCH_SOURCES to look in kernel dirs just for that one, and not of other parts. Then there is no ambiguity.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>