Update patch set 1
Patch Set 1:

> So there are cases where custom list is still required?

At the present time there are many cases.  Where it makes sense and is possible (as in this PR) I will replace those with standard collections from Haiku's standard classes.  After that I need to decide what to do with the List class.  I can see that some improvements might be possible and it could continue to be used or I could alternatively replace its use with STL collections instead if it is possible to do this with the older libs in the gcc2 build.  The advantage of using the STL collections is that there would be less surface area to maintain.

Patch-set: 1
Reviewer: Gerrit User 1000018 <1000018@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>