Update patch set 2
Patch Set 2:

> Got it, I'm so used to working in branches where future commits in the branch get stacked automatically when you do a later push, gerrit making a new ticket confused me.  If doing multiple commits, does every commit after the ticket is made, need to have the Change ID? and does every commit message need to have "component: commit message" in it, even if they're going to get all submitted together?

No, if you want to submit a branch with multiple commits, each commit will be reviewed separately (and each gets a separate change id). Once all are reviewed they can be submitted (the equivalent of merging, in the branch flow) all at the same time.

It is usually easier to review changes by small pieces this way rather than by large pull requests. The link between the commits is not as directly visible as in pull-request based tools, but is still tracked by the "relation chain" and "submitted together" lists, on the top right of the change overview page.


The "component:" thing is just a convention from Haiku devs, and not consistently followed I think. It's just useful to know easily what a commit is referring to.

Patch-set: 2