Update patch set 5
Patch Set 5:

This should be it! Everything looks and feels a lot more cleaner (almost like the original from BeOS!)...though this may create a mess for decorators out-of-tree which will need to be rebuilt and modified to not override certain methods. Honestly, we need a stable "BDecorator" API that would allow us to provide a clean, single API for decorator builders to use, and not have to worry about internal changes in the app_server. (And also have lots of FBC padding)

Patch-set: 5
Reviewer: Gerrit User 1000052 <1000052@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>