Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > Yeah but are you okay with the O(n) here? The loop we are trying to remove would do it in O(logn)....
> 
> Now I don't understand.
> 
> The loop implements a binary search. It can be implemented with bsearch with the same complexity.
> 
> The second loop (which is already there) will be O(n) no matter what, but with a small n (the number of items having the same hash, if I follow correctly). That's where the theoretical complexity approximation breaks up. An O(logn) algorithm is nice, but if it's in fact 10000 * log(n), and you have another algorithm in 10 * n, for small values of n the second algorithm is actually faster.
> 
> Do I miss something? Is this loop not in fact a binary search? It looks very much like one to me.

The first loop is a binary search but it finds the first instance of the hash value. Using bsearch() we won't find that as it could return a pointer to any of the instances. Looking backwards from this pointer to find the first instance (this will take O(n)), kind of seems unnecessary to me.
You're right about the second loops complexity and how overall it is O(n). I hope this makes sense.

Patch-set: 8
Reviewer: Gerrit User 1000183 <1000183@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>