Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > The first loop is a binary search but it finds the first instance of the hash value. Using bsearch() we won't find that as it could return a pointer to any of the instances. Looking backwards from this pointer to find the first instance (this will take O(n)), kind of seems unnecessary to me.
> > You're right about the second loops complexity and how overall it is O(n). I hope this makes sense.
> 
> Ok now I see the difference from a typical binary search in not stopping when an equal key is found. That's a subtle but important difference and I think deserves a comment explaining it (there is one already but it wasn't very clear to me)

Done. Also, if I may suggest, should I rename this from Extent to Block?
Extent really consists of leaf and block directories and many things differ but some things are similar. I was thinking of having a different implementation for leaf. It'll be easier to deal with.

Patch-set: 8