Update patch set 1
Patch Set 1:

You're cheating, the two screenshots were taken with different font sizes ;)

New position looks better, but I'm not sure the hardcoded 3 is the way to go. I think what we want is the box border exactly aligned with the text baseline, no matter which font and font sizes are used?

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>