WIP: Update the Input_Server files to get the settings working
I try to print the name of the mouse in the logs. I write the code in the
InterfaceDefs.cpp to Add the name and the Find that name. And these are getting
printed in the console. but, when I write them in the InputServer.cpp
and trying to print them in the logs, I get the names in the binary format.

Change-Id: I3d9880935990e671e6f98ff1895edc1e448000e0