WIP: Update the Input_Server files to get the settings working
I have made the different functions in the InterfaceDefs.h and used them in the MouseSettings.h and MouseInputDevice.g files.
Now, the code runs but there is not effect on the settings. The settings are workings defaults.
Because the MouseSettings.h files linkes with the src/servers/input/InputServer.h file.

- The logs are not get printed in the consolve taht I have written in the InputServer.h file.

Change-Id: I3d9880935990e671e6f98ff1895edc1e448000e0