WIP: Update the Input_Server files to get the settings working
I made some changes in the code again and try to write the logs cleanly and
erase the extra logs. Now, when I run the input preferences and try to change
the mouse type, the debugger window opens, after killing it. The mouse hangs
and stop working.
Please have a review if you want to see the logs then, just run the command
"cat /boot/system/var/log/syslog | grep -n "DEBUG_MOUSE". It will give you all
the logs with the file name which I have added to debug and test the code also
to understand the flow of code.
Please help me if anybody have any idea where I am worng and how to make
this work. I have tried every logic that came to my mind but for a long time I
got stuck in this and now, no clue how to finish this work! :(

Change-Id: I3d9880935990e671e6f98ff1895edc1e448000e0