Update patch set 5
Patch Set 5:

> Patch Set 2:
> 
> The alignment was computed previously so that when you use a BPopUpMenu as the box header (Background prefs in Haiku does this, for example), the text is aligned at the same place. We should check that it's still the case (moving the BPopUpMenu up as needed to match the new text alignment)

My goal here is to reproduce what was happening on BeOS R5 for maximum compatibility. Here is a screenshot of the Backgrounds app on BeOS R5:
https://64.media.tumblr.com/7cb05a64bca55399480c32ffb461c07c/e5107c9f7ba94bc9-d2/s500x750/4dfdd88aa8891781adffaadb4f91543368a1aab6.png

And here is a screenshot of the same BeOS R5 backgrounds app running on Haiku:
https://64.media.tumblr.com/b3ca61aa7a12907acc1a870ec5f6724d/e5107c9f7ba94bc9-cf/s500x750/d6159340d0dd9edca0ad076fbae1fe1017cdf663.png

Notice in these screenshots that the dropdown is centered in the box while the text is slightly above center. This is what I'm trying to achieve on Haiku to fix issues with old BeOS apps that have text label truncated. If I were to move the menu field up it would not match BeOS behavior and get truncated by apps that expect the BeOS position. What I'm doing is intentionally different than what you are proposing.

Patch-set: 5
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>