Update patch set 2
Patch Set 2:

The changes generally look sane, but the main problem as discussed on IRC is that we're missing the appropriate DWARF register numbers for ymm and/or zmm in order to unwind those correctly from register rules. However, we haven't yet managed to locate a version of the AMD64 ELF ABI that's new enough to include zmm, so that may need to be found in gcc and/or binutils source directly.

Patch-set: 2
CC: Gerrit User 1000015 <1000015@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>