Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > #15189
> 
> Ticket 15189 does not exist.

It was deleted, but it existed before July 20th. It was reported by telephil and was entitled "BTextControl right alignment error" It had a nice test program that created a window with a right-aligned text control that demonstrated the bug. But after commenting on the ticket the other day it was suddenly deleted. If anybody knows how to undelete a ticket I'd appreciate #15189 being brought back.

Patch-set: 2
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>