Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I think this is a problem for languages not using "arabic" numbers since the limitation will be in bytes.
> 
> Would SetMaxBytes(4) be better?

According to https://www.fileformat.info/info/unicode/category/Nd/list.htm some of the digits (at the end of the list) require 4 bytes each in utf-8. Apparently Adlam is in active use in West Africa.

> 
> > I'd rather let the user enter anything and use MarkAsInvalid() to hint at problems (since the limit is not at 99 worspaces anyway?)
> 
> The workspaces limit is 32. MarkAsInvalid() seems to only exist on BTextControl, Unfortunalely I only have access to the BTextView. Shouldn't MarkAsInvalid() be moved to BTextView and BTextControl should call it?

I don't know if it is useful as a public API in BTextView? It can be done this way or adding a similar method to BSpinner (if it can be done in BTextControl, it can be done in BSpinner in the same way)

Patch-set: 1