Update patch set 4
Patch Set 4:

(1 comment)

So I have rewritten this patch to store settings in the workspaces settings as it should be done (alongside the video mode and other app_server settings). However, the attempt to save it on shutdown doesn't work. Should this method be deleted if it's never called? Or is the problem elsewhere?

Patch-set: 4
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>