Update patch set 4
Patch Set 4:

> > > How do I fix for 64 bit and should I also subtract another 4 bytes for LayoutData* fLayoutData pointer which was also added after BeOS R5. I checked BeOS R5 and the sizeof BTextView was 356 bytes. On 32 bit Haiku the size remains 356 bytes after patch set 3. The sizeof BTextView is 512 bytes on 64bit before my changes. Do we care about the class size on 64-bit? Is there padding on class size on 64bit?
> > 
> > Something like this is possible:
> > 
> > uint32 _reserved[6 - sizeof(BRect*)/sizeof(uint32)];
> 
> Shouldn’t that be done for all pointers on the class? That would put us in the negative for reserved size...

No, it should just preserve the current size. It's normal that the 64bit objects are larger than the 32bit ones, and it does not matter. But when we add a field, we should do things so that it does not change the size on either architecture, so that compatibility with existing apps is preserved. And yes, this means we may run out of free space earlier on 64bit.

Do we care? I'd say yes, now that we have shipped two beta versions with 64bit support. It would not be nice to break apps there.

Actually this will become more and more complicated as we add more architectures (as you may need to take padding into account as well, and that's platform dependant).

Hopefully we can do something better for R2 (C++ modules? moving the data to a separate place so the user-visible class only has a single field which is a pointer to the data?), but for now we have to deal with what Be could do and how we managed it since then (probably we should have put larger padding for 64bit when we made it an official ABI, but no one thought of it).

Patch-set: 4
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>